+++ docs/src/examples/reactivity/helpers/reactive-get-index/index.js
@@ -1,15 +1,15 @@
 // Helper: getIndex
-import { Reaction, Signal } from '@semantic-ui/reactivity';
+import { flush, reaction, signal } from '@semantic-ui/reactivity';
 
-const colors = new Signal(['red', 'green', 'blue', 'yellow']);
+const colors = signal(['red', 'green', 'blue', 'yellow']);
 
-Reaction.create(() => console.log('First color:', colors.getIndex(0)));
-Reaction.create(() => console.log('Third color:', colors.getIndex(2)));
+reaction(() => console.log('First color:', colors.getIndex(0)));
+reaction(() => console.log('Third color:', colors.getIndex(2)));
 
 // Change the first color
 colors.setIndex(0, 'purple');
-Reaction.flush();
+flush();
 
 // Change the third color
 colors.setIndex(2, 'orange');
-Reaction.flush();
+flush();