+++ docs/src/examples/reactivity/helpers/reactive-set-property/index.js
@@ -1,15 +1,15 @@
-import { Reaction, Signal } from '@semantic-ui/reactivity';
+import { reaction, signal } from '@semantic-ui/reactivity';
 
-const users = new Signal([
+const users = signal([
   { id: 'aliceuser', name: 'Alice', health: 'Bad' },
   { id: 'samuser', name: 'Sam', health: 'Okay' },
 ]);
 
 // Log any changes in users
-Reaction.create(() => console.log(users.value));
+reaction(() => console.log(users.value));
 
 // changes aliceuser's name to 'Allison'
-users.setProperty('aliceuser', 'name', 'Allison');
+users.setItemProperty('aliceuser', 'name', 'Allison');
 
 // changes everyone's health to 'Good'
-users.setArrayProperty('health', 'Good');
+users.setProperty('health', 'Good');